# Conflicts:
#	Online_Shop/Web.config

+++ Online_Shop/Areas/Admin/Controllers/UserController.cs
@@ -1,6 +1,7 @@
 ﻿using Online_Shop.Controllers;
 using Online_Shop.Models;
 using Online_Shop.Models.DTO;
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -28,37 +29,62 @@ namespace Online_Shop.Areas.Admin.Controllers
 
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create(UserRoles userRole, HttpPostedFileBase file)
+        public ActionResult Create(UserRoles userRoles, HttpPostedFileBase file)
         {
             if (ModelState.IsValid)
             {
-                if (db.Users.Any(a => a.Name == userRole.User.Name))
+                if (userRoles.User.User_name.Contains(' '))
+                {
+                    ModelState.AddModelError("User_name", "This name must have no space ");
+                    return View();
+                }
+                if (db.Users.Any(m => m.User_name == userRoles.User.User_name))
+                {
+                    ModelState.AddModelError("User_name", "This name already existed in the Database ");
+                    return View();
+                }
+                if (userRoles.User.Password == null)
+                {
+                    ModelState.AddModelError("Password", "The password is required");
+                    return View();
+                }
+                //if (!userRoles.User.Password.Equals(userRoles.User.RePassword))
+                //{
+                //    ModelState.AddModelError("User_name", "This name must have no space ");
+                //    return View();
+                //}
+                if (db.Users.Any(a => a.Name == userRoles.User.Name))
                 {
                     ModelState.AddModelError("Name", "This name already existed in the Database");
                     return View();
                 }
-
+                if (DateTime.Compare(userRoles.User.Dob, DateTime.Now) > 0)
+                {
+                    ModelState.AddModelError("Dob", "The date of birth is invalid");
+                    return View();
+                }
                 if (file != null)
                 {
 
-                    string InputFileName = userRole.User.Id + Path.GetFileName(file.FileName);
+                    string InputFileName = userRoles.User.Id + Path.GetFileName(file.FileName);
                     string ServerSavePath = Path.Combine(Server.MapPath("~/Include/Images/"), InputFileName);
                     file.SaveAs(ServerSavePath);
 
-                    userRole.User.Avatar = InputFileName;
+                    userRoles.User.Avatar = InputFileName;
                 }
                 else
                 {
-                    userRole.User.Avatar = "";
+                    userRoles.User.Avatar = "";
                 }
-                userRole.User.Password = BCrypt.Net.BCrypt.HashPassword(userRole.User.Password);
-                db.Users.Add(userRole.User);
+                userRoles.User.Password = BCrypt.Net.BCrypt.HashPassword(userRoles.User.Password);
+                userRoles.User.RePassword = userRoles.User.Password;
+                db.Users.Add(userRoles.User);
                 db.SaveChanges();
                 TempData["Status"] = "User Created Successfully";
                 return RedirectToAction("Index");
             }
-            userRole.Roles = db.Roles.ToList();
-            return View();
+            userRoles.Roles = db.Roles.ToList();
+            return View(userRoles);
         }
 
         public ActionResult Edit(int? id)
@@ -79,13 +105,24 @@ namespace Online_Shop.Areas.Admin.Controllers
         {
             if (ModelState.IsValid)
             {
+                if (DateTime.Compare(userRoles.User.Dob, DateTime.Now) > 0)
+                {
+                    ModelState.AddModelError("Dob", "The date of birth is invalid");
+                    userRoles.Roles = db.Roles.ToList();
+                    return View(userRoles);
+                }
                 User checkUser = db.Users.FirstOrDefault(u => u.User_name == userRoles.User.User_name);
                 if (checkUser == null || checkUser.Id == userRoles.User.Id)
                 {
                     User user = db.Users.Find(userRoles.User.Id);
                     if (file != null)
                     {
-                        System.IO.File.Delete(Path.Combine(Server.MapPath("~/Include/Images"), user.Avatar));
+                        if (user.Avatar != "")
+                        {
+                            System.IO.File.Delete(Path.Combine(Server.MapPath("~/Include/Images"), user.Avatar));
+                        }
+
+
                         string InputFileName = Path.GetFileName(file.FileName);
                         string ServerSavePath = Path.Combine(Server.MapPath("~/Include/Images/"), InputFileName);
                         file.SaveAs(ServerSavePath);
@@ -97,6 +134,9 @@ namespace Online_Shop.Areas.Admin.Controllers
                     {
                         userRoles.User.Avatar = user.Avatar;
                     }
+                    User u = db.Users.Find(userRoles.User.Id);
+                    userRoles.User.Password = u.Password;
+                    userRoles.User.RePassword = userRoles.User.Password;
                     db.Entry(db.Users.Find(userRoles.User.Id))
                         .CurrentValues.SetValues(userRoles.User);
                     db.SaveChanges();
@@ -112,55 +152,65 @@ namespace Online_Shop.Areas.Admin.Controllers
             userRoles.Roles = db.Roles.ToList();
             return View(userRoles);
         }
+
+        //public ActionResult UpdatePassword(int? id)
+        //{
+
+        //    if (!db.Users.Any(u => u.Id == id) || id == null)
+        //    {
+        //        return RedirectToAction("Index", "Dashboard");
+        //    }
+        //    User user = db.Users.Find(id);
+        //    Password p = new Password()
+        //    {
+        //        UserId = (int)id,
+        //        CurrentPassword = user.Password,
+        //        UserName = user.User_name
+        //    };
+        //    return View(p);
+        //}
+
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult UpdatePassword(int id)
+        [HandleError]
+        public ActionResult UpdatePassword(int? id, Password p)
         {
+            if (id == null)
+            {
+                return RedirectToAction("Index", "Dashboard");
+            }
+            if (ModelState.IsValid)
+            {
+
+                if (!p.NewPassword.Equals(p.ReNewPassword))
+                {
+                    return PartialView("_Password", p);
+                }
+                User u = db.Users.Find(id);
+                u.Password = BCrypt.Net.BCrypt.HashPassword(p.NewPassword);
+                u.RePassword = db.Users.Find(id).Password;
+                db.SaveChanges();
+                ViewBag.Message = "Password Updated";
+                ModelState.Clear();
+                return PartialView("_Password", p);
+
+            }
+            return PartialView("_Password", p);
+        }
+
+        [HttpPost]
+
+        public ActionResult ChangeRole(int id, FormCollection form)
+        {
+
             if (!db.Users.Any(u => u.Id == id))
             {
                 return RedirectToAction("Index", "Dashboard");
             }
-            Password p = new Password()
-            {
-                UserId = id,
-                CurrentPassword = db.Users.Find(id).Password
-            };
-            return View(p);
-        }
-
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public ActionResult SavePassword(Password p)
-        {
-            if (ModelState.IsValid)
-            {
-                if (!BCrypt.Net.BCrypt.Verify(p.CurrentPasswordInput, p.CurrentPassword))
-                {
-                    ModelState.AddModelError("CurrentPasswordInput", "Please enter the correct current password");
-                    return View(p);
-                }
-                if (!p.NewPassword.Equals(p.ReNewPassword))
-                {
-                    ModelState.AddModelError("NewPassword", "Two passwords must be the same");
-                    return View(p);
-                }
-                db.Users.Find(p.UserId).Password = BCrypt.Net.BCrypt.HashPassword(p.NewPassword);
-                db.SaveChanges();
-                return RedirectToAction("Index");
-            }
-            return View();
-        }
-
-        [HttpPost]
-        public ActionResult ChangeRole(int userId, int newRole_id)
-        {
-            if (!db.Users.Any(u => u.Id == userId))
-            {
-                return RedirectToAction("Index", "Dashboard");
-            }
-            db.Users.Find(userId).Role_id = newRole_id;
+            db.Users.Find(id).Role_id = Convert.ToInt32(form["Role"]);
             db.SaveChanges();
-            return Json(new { success = true, message = "Role Updated" }, JsonRequestBehavior.AllowGet);
+            //return Json(new { success = true, message = "Role Updated" }, JsonRequestBehavior.AllowGet);
+            return RedirectToAction("Index");
         }
 
         [HttpPost]
